Use bats to test basic Nova tasks on OpenStack system
This commit modifies the .kitchen.yml file to use the proxy driver
to run against a system idenfied by the environment variable `SUT_HOST`
that has OpenStack deployed to it.

A bats test for basic Nova operations is included.

The README has been updated to provide a full instruction set on how to
set up the test runner environment to use this repo. The setup of the
external OpenStack system is not covered here.